StreamingCmdlet<byte[], Integer>
This might be useful if/when I try to allow
commands to be composed, somehow.

The actual motivation was simply to reduce
duplication between different sub-commands
that all work about the same way, emitting
a bunch of things in a configurable format
before returning some exit code.

WalkFilesystemCmd still emits some stuff directly
to stdout and stderr, which should be fixed.